move memdebug declarations to curl_setup.h, drop a macro hack
